refatorando função de logar com google, separando responsabilidades
